refactor: changed directory structure, bump version
